Support for apps without shadow root (WIP).
