+++ .changeset/beige-hairs-tease.md
@@ -1,0 +1,7 @@
+---
+"@open-pioneer/runtime": major
+---
+
+Support running an application without using a ShadowRoot.
+
+TODO: Describe pro/con
