+++ src/packages/runtime/react-integration/ErrorScreen.tsx
@@ -1,7 +1,7 @@
 // SPDX-FileCopyrightText: 2023-2025 Open Pioneer project (https://github.com/open-pioneer)
 // SPDX-License-Identifier: Apache-2.0
 import { Alert, Box, Heading, Text, VStack } from "@chakra-ui/react";
-import { PackageIntl } from "./i18n";
+import { PackageIntl } from "../i18n";
 
 const MESSAGES_DE = {
     "title": "Anwendungsstart fehlgeschlagen",
